Move most AI logic to new chunk #951
