+++ browser/data-browser/package.json
@@ -44,7 +44,6 @@
     "ollama-ai-provider": "^1.2.0",
     "polished": "^4.3.1",
     "prismjs": "^1.29.0",
-    "query-string": "^7.1.3",
     "quick-score": "^0.2.0",
     "react": "^19.0.0",
     "react-colorful": "^5.6.1",
