+++ browser/data-browser/src/chunks/AI/AIChatMessageParts/AssistantMessage.tsx
