+++ browser/data-browser/src/chunks/AI/AIChatMessageParts/UserMessage.tsx
@@ -1,7 +1,7 @@
 import { type AtomicUIMessage } from '../types';
 import { styled } from 'styled-components';
-import Markdown from '../../datatypes/Markdown';
-import { Row } from '../../Row';
+import Markdown from '@components/datatypes/Markdown';
+import { Row } from '@components/Row';
 import { MessageContextItem } from '../MessageContextItem';
 import { FileContent } from './FileContent';
