+++ browser/data-browser/src/chunks/AI/ClientOnlyTransport.ts
