+++ browser/data-browser/src/chunks/AI/atomicSchemaHelpers.ts
