+++ browser/data-browser/src/components/ResourceContextMenu/index.tsx
@@ -40,10 +40,7 @@ import { useNewRoute } from '../../helpers/useNewRoute';
 import { addIf } from '../../helpers/addIf';
 import { useNavigateWithTransition } from '../../hooks/useNavigateWithTransition';
 import { newContextItem, useAISidebar } from '../AI/AISidebarContext';
-import {
-  isAtomicResourceContext,
-  type AIAtomicResourceMessageContext,
-} from '../AI/types';
+import { type AIAtomicResourceMessageContext } from '@chunks/AI/types';
 
 export const ContextMenuOptions = {
   View: 'view',
@@ -105,7 +102,9 @@ export function ResourceContextMenu({
 
   const addToChat = () => {
     setContextItems(prev => [
-      ...prev.filter(x => isAtomicResourceContext(x) && x.subject !== subject),
+      ...prev.filter(
+        x => x.type === 'atomic-resource' && x.subject !== subject,
+      ),
       newContextItem({
         type: 'atomic-resource',
         subject,
